pytorch-neural-style-transfer-master/requirements.txt
matplotlib>=3.8.0
torch>=2.0
torchvision>=0.15
numpy>=1.21.0,<1.26.0
opencv-python>=4.7
streamlit>=1.28.0
Pillow>=9.0.0
tqdm>=4.64.0